On branch main  Changes to be committed: 	modified:   examples/simple_inference/README.md 	deleted:    examples/simple_inference/images/result.jpg 	modified:   examples/simple_inference/main.rs 	new file:   models/README.MD 	renamed:    export_torchscript.py -> models/export_torchscript.py 	new file:   models/yolo.torchscript 	new file:   models/yolov5s.pt 	modified:   src/yolo.rs
Added export script. Strucured repo for better examples.
Added more detail.
