+++ examples/simple_inference/README.md
@@ -1,31 +1,20 @@
 
-This example demonstrates how to run the yolov5 inference.
+This example demonstrates how to run the Yolov5 inference.
 
-First, you have to create an export of weights in torchscript format. This can be done using the https://github.com/ultralytics/yolov5 repository by running the command:
-`python export.py --imgsz 384 640`
+1. Clone this repository to the desired location on you computer.
+    >> git clone 
 
-Then you need to get rid of the tuple format on the model output, you can do this with a script:
-```
-import torch
-import torch.nn as nn
+2. Build and run the example using:  
 
+    >> cd rusty/yolo/examples/simple_inference
 
-class TorchScriptWithoutTuple(nn.Module):
-    def __init__(self, model) -> None:
-        super().__init__()
-        self.model = model
+    >> cargo run 
 
-    def forward(self, x):
-        return self.model(x)[0]
+3. This will perform inference on the image defined in line 15 of [main.rs](main.rs), using the Yolo torchscrpt file defined  in line 7.
 
-def main():
-    model = torch.hub.load('ultralytics/yolov5', 'custom', "yolov5s.torchscript").eval()
-    save_model = TorchScriptWithoutTuple(model)
-    traced_model = torch.jit.trace(save_model, (torch.rand((1, 3, 384, 640))))
-    torch.jit.save(traced_model, "weights/model.pt",)
+4. The result will be output to the [images](images/) folder once complete, as shown on line 25 of [main.rs](main.rs).
 
 
-if __name__ == "__main__":
-    main()
+For convenience, a copy a the Yolov5s model in torchscript format has been proivided in the [models](../../models/) folder. By deault this is where the example in [main.rs](main.rs) will look for the model. This can be changed by defining the path to a different Yolo torchscript file in the *"weights"* variable in line 7 of [main.rs](main.rs).
 
-```
+Several different versions of Yolov5 are available, with varying degrees of speed and accuracy. The TCH crate expects a slightly different format to what is provided by Yolo's built in export funtion. As such, please follow the **instructions** provided [here](../../models/) to obtain torchscrpt files for other versions of yolo.
