+++ src/yolo.rs
@@ -70,8 +70,7 @@ impl YOLO {
 
         let pred = self
             .model
-            .forward_ts(&[img])
-            .unwrap()
+            .forward_ts(&[img]).unwrap()
             .to_device(tch::Device::Cpu);
         let result = self.non_max_suppression(&pred.get(0), conf_thresh, iou_thresh);
